move dropdown to components
